chore: apply pr suggestions
