+++ packages/shared/lib/core/layer-2/utils/index.ts
@@ -2,12 +2,11 @@ export * from './addGasBudget'
 export * from './evmAddressToAgentID'
 export * from './getAbiForContractType'
 export * from './getAgentBalanceParameters'
-export * from './getErc20MetadataFromContract'
 export * from './getDestinationNetworkFromAddress'
+export * from './getErc20TokenMetadata'
 export * from './getLayer2MetadataForTransfer'
 export * from './getLayer2NetworkFromAddress'
 export * from './getSmartContractHexName'
-export * from './getErc20TokenSymbol'
 export * from './parseLayer2Metadata'
 export * from './parseLayer2MetadataForTransfer'
 export * from './prepareErc20EvmTransactionData'
